fix: circular dependency, move program and project from common package into data package
