+++ libs/common/src/index.ts
@@ -5,8 +5,6 @@ export * from './establishment/validator/siretValidator'
 export * from './establishment/validator/emailValidator'
 export * from './establishment/validator/phoneValidator'
 export * from './opportunity/opportunityTypes'
-export { default as Program } from './program/program'
-export * from './program/types'
 export * from './questionnaire/types/questionnaireData'
 export * from './questionnaire/types/types'
 export * from './questionnaire/objectiveChecker'
