+++ libs/data/src/index.ts
@@ -1,3 +1,3 @@
-export * from './type/program'
+export * from './program/program'
 export * from './type/publicodesTypes'
 export * from './generated/program'
