+++ libs/data/src/program/program.ts
@@ -1,5 +1,4 @@
 import { ProgramAidType } from './types'
-import { ProgramType } from '@tee/data'
 
 export default class Program {
   public static getPrefixedProgramName(program: ProgramType) {
@@ -21,3 +20,8 @@ export default class Program {
     return `${prefix}${program["nature de l'aide"]} ${program.titre}`
   }
 }
+
+import type { Dispositif as ProgramWithoutId } from '../generated/program'
+
+export type { ProgramWithoutId }
+export type ProgramType = ProgramWithoutId & { id: string }
