+++ libs/data/src/projects/type/index.ts
@@ -1,0 +1,2 @@
+export * from './project'
+export * from './baserow'
