--- libs/data/src/type/program.ts
@@ -1,4 +1,0 @@
-import type { Dispositif as ProgramWithoutId } from '../generated/program'
-
-export type { ProgramWithoutId }
-export type ProgramType = ProgramWithoutId & { id: string }
