+++ libs/data/static/index.ts
@@ -1,0 +1,1 @@
+export { default as projects } from './projects.json'
