+++ tsconfig.base.json
@@ -22,6 +22,7 @@
       "@tee/common": ["libs/common/src/index.ts"],
       "@tee/data": ["libs/data/src/index.ts"],
       "@tee/data/references": ["libs/data/references/index.ts"],
+      "@tee/data/static": ["libs/data/static/index.ts"],
       "@tee/data/generated": ["libs/data/generated/index.ts"]
     }
   },
